Refactor modules to shrink overall module size
